+++ core/View.php
@@ -87,6 +87,7 @@ if (!defined('PIWIK_USER_PATH')) {
  * - **isPluginLoaded**: Returns true if the supplied plugin is loaded, false if otherwise.
  *                       `{% if isPluginLoaded('Goals') %}...{% endif %}`
  * - **arePiwikProAdsEnabled**: Returns true if it is ok to show some Piwik PRO advertising in the UI (from Piwik 2.16.0)
+ * - **isMultiServerEnvironment**: Returns true if Piwik is used on more than one server (from 2.16.1)
  *
  * ### Examples
  *
@@ -228,6 +229,7 @@ class View implements ViewInterface
             $this->latest_version_available = UpdateCheck::isNewestVersionAvailable();
             $this->disableLink = Common::getRequestVar('disableLink', 0, 'int');
             $this->isWidget = Common::getRequestVar('widget', 0, 'int');
+            $this->isMultiServerEnvironment = SettingsPiwik::isMultiServerEnvironment();
 
             $piwikProAds = StaticContainer::get('Piwik\PiwikPro\Advertising');
             $this->arePiwikProAdsEnabled = $piwikProAds->arePiwikProAdsEnabled();