+++ plugins/CoreHome/Controller.php
@@ -234,7 +234,7 @@ class Controller extends \Piwik\Plugin\Controller
         // perform check (but only once every 10s)
         UpdateCheck::check($force = false, UpdateCheck::UI_CLICK_CHECK_INTERVAL);
 
-        $marketplace = StaticContainer::get('Piwik\Plugins\Marketplace\MarketplaceApi');
+        $marketplace = StaticContainer::get('Piwik\Plugins\Marketplace\Api\Client');
         $marketplace->clearAllCacheEntries();
 
         $view = new View('@CoreHome/checkForUpdates');