+++ plugins/CorePluginsAdmin/Controller.php
@@ -50,11 +50,17 @@ class Controller extends Plugin\ControllerAdmin
      */
     private $marketplacePlugins;
 
+    /**
+     * @var Plugin\Manager
+     */
+    private $pluginManager;
+
     public function __construct(Translator $translator, PluginInstaller $pluginInstaller, Plugins $marketplacePlugins)
     {
         $this->translator = $translator;
         $this->pluginInstaller = $pluginInstaller;
         $this->marketplacePlugins = $marketplacePlugins;
+        $this->pluginManager = Plugin\Manager::getInstance();
 
         parent::__construct();
     }
@@ -135,7 +141,7 @@ class Controller extends Plugin\ControllerAdmin
             'name'        => $pluginMetadata->name,
             'version'     => $pluginMetadata->version,
             'isTheme'     => !empty($pluginMetadata->theme),
-            'isActivated' => \Piwik\Plugin\Manager::getInstance()->isPluginActivated($pluginMetadata->name)
+            'isActivated' => $this->pluginManager->isPluginActivated($pluginMetadata->name)
         );
 
         return $view->render();
@@ -183,17 +189,15 @@ class Controller extends Plugin\ControllerAdmin
         $view->activateNonce = Nonce::getNonce(static::ACTIVATE_NONCE);
         $view->uninstallNonce = Nonce::getNonce(static::UNINSTALL_NONCE);
         $view->deactivateNonce = Nonce::getNonce(static::DEACTIVATE_NONCE);
-        $view->installNonce = Nonce::getNonce(static::INSTALL_NONCE);
         $view->pluginsInfo = $this->getPluginsInfo($themesOnly);
 
         $users = Request::processRequest('UsersManager.getUsers');
         $view->otherUsersCount = count($users) - 1;
-        $view->themeEnabled = \Piwik\Plugin\Manager::getInstance()->getThemeEnabled()->getPluginName();
+        $view->themeEnabled = $this->pluginManager->getThemeEnabled()->getPluginName();
 
         $view->pluginNamesHavingSettings = $this->getPluginNamesHavingSettingsForCurrentUser();
         $view->isMarketplaceEnabled = Marketplace::isMarketplaceEnabled();
         $view->isPluginsAdminEnabled = CorePluginsAdmin::isPluginsAdminEnabled();
-        $view->isMultiServerEnvironment = SettingsPiwik::isMultiServerEnvironment();
 
         $view->pluginsHavingUpdate    = array();
         $view->marketplacePluginNames = array();
@@ -201,10 +205,7 @@ class Controller extends Plugin\ControllerAdmin
         if (Marketplace::isMarketplaceEnabled()) {
             try {
                 $view->marketplacePluginNames = $this->marketplacePlugins->getAvailablePluginNames($themesOnly);
-
-                $pluginsHavingUpdate = $this->marketplacePlugins->getPluginsHavingUpdate(true);
-                $themesHavingUpdate  = $this->marketplacePlugins->getPluginsHavingUpdate(false);
-                $view->pluginsHavingUpdate    = $pluginsHavingUpdate + $themesHavingUpdate;
+                $view->pluginsHavingUpdate    = $this->marketplacePlugins->getPluginsHavingUpdate();
             } catch(Exception $e) {
                 // curl exec connection error (ie. server not connected to internet)
             }
@@ -244,12 +245,11 @@ class Controller extends Plugin\ControllerAdmin
 
     protected function getPluginsInfo($themesOnly = false)
     {
-        $pluginManager = \Piwik\Plugin\Manager::getInstance();
-        $plugins = $pluginManager->loadAllPluginsAndGetTheirInfo();
+        $plugins = $this->pluginManager->loadAllPluginsAndGetTheirInfo();
 
         foreach ($plugins as $pluginName => &$plugin) {
 
-            $plugin['isCorePlugin'] = $pluginManager->isPluginBundledWithCore($pluginName);
+            $plugin['isCorePlugin'] = $this->pluginManager->isPluginBundledWithCore($pluginName);
 
             if (!empty($plugin['info']['description'])) {
                 $plugin['info']['description'] = $this->translator->translate($plugin['info']['description']);
@@ -336,7 +336,7 @@ class Controller extends Plugin\ControllerAdmin
         $view->lastError   = $lastError;
         $view->isSuperUser = Piwik::hasUserSuperUserAccess();
         $view->isAnonymousUser = Piwik::isUserIsAnonymous();
-        $view->plugins         = Plugin\Manager::getInstance()->loadAllPluginsAndGetTheirInfo();
+        $view->plugins         = $this->pluginManager->loadAllPluginsAndGetTheirInfo();
         $view->deactivateNonce = Nonce::getNonce(static::DEACTIVATE_NONCE);
         $view->uninstallNonce  = Nonce::getNonce(static::UNINSTALL_NONCE);
         $view->emailSuperUser  = implode(',', Piwik::getAllSuperUserAccessEmailAddresses());
@@ -360,16 +360,9 @@ class Controller extends Plugin\ControllerAdmin
         $pluginName = $this->initPluginModification(static::ACTIVATE_NONCE);
         $this->dieIfPluginsAdminIsDisabled();
 
-        \Piwik\Plugin\Manager::getInstance()->activatePlugin($pluginName);
+        $this->pluginManager->activatePlugin($pluginName);
 
         if ($redirectAfter) {
-            $plugin = \Piwik\Plugin\Manager::getInstance()->loadPlugin($pluginName);
-
-            $actionToRedirect = 'plugins';
-            if ($plugin->isTheme()) {
-                $actionToRedirect = 'themes';
-            }
-
             $message = $this->translator->translate('CorePluginsAdmin_SuccessfullyActicated', array($pluginName));
             if (SettingsManager::hasSystemPluginSettingsForCurrentUser($pluginName)) {
                 $target   = sprintf('<a href="index.php%s#%s">',
@@ -384,7 +377,7 @@ class Controller extends Plugin\ControllerAdmin
             $notification->context = Notification::CONTEXT_SUCCESS;
             Notification\Manager::notify('CorePluginsAdmin_PluginActivated', $notification);
 
-            $this->redirectToIndex('CorePluginsAdmin', $actionToRedirect);
+            $this->redirectAfterModification($redirectAfter);
         }
     }
 
@@ -393,7 +386,7 @@ class Controller extends Plugin\ControllerAdmin
         $pluginName = $this->initPluginModification(static::DEACTIVATE_NONCE);
         $this->dieIfPluginsAdminIsDisabled();
 
-        \Piwik\Plugin\Manager::getInstance()->deactivatePlugin($pluginName);
+        $this->pluginManager->deactivatePlugin($pluginName);
         $this->redirectAfterModification($redirectAfter);
     }
 
@@ -402,7 +395,7 @@ class Controller extends Plugin\ControllerAdmin
         $pluginName = $this->initPluginModification(static::UNINSTALL_NONCE);
         $this->dieIfPluginsAdminIsDisabled();
 
-        $uninstalled = \Piwik\Plugin\Manager::getInstance()->uninstallPlugin($pluginName);
+        $uninstalled = $this->pluginManager->uninstallPlugin($pluginName);
 
         if (!$uninstalled) {
             $path = Filesystem::getPathToPiwikRoot() . '/plugins/' . $pluginName . '/';