+++ plugins/Marketplace/Marketplace.php
@@ -8,8 +8,13 @@
 
 namespace Piwik\Plugins\Marketplace;
 
+use Piwik\API\Request;
+use Piwik\Cache;
 use Piwik\Config;
+use Piwik\Container\StaticContainer;
+use Piwik\Piwik;
 use Piwik\Plugin;
+use Exception;
 
 class Marketplace extends \Piwik\Plugin
 {
@@ -22,7 +27,30 @@ class Marketplace extends \Piwik\Plugin
             'AssetManager.getJavaScriptFiles' => 'getJsFiles',
             'AssetManager.getStylesheetFiles' => 'getStylesheetFiles',
             'Translate.getClientSideTranslationKeys' => 'getClientSideTranslationKeys',
+            'Request.getRenamedModuleAndAction' => 'getRenamedModuleAndAction',
         );
+    }
+
+    public function getRenamedModuleAndAction(&$module, &$action)
+    {
+        $pluginManager = self::getPluginManager();
+
+        if ($pluginManager->isPluginBundledWithCore($module)) {
+            // make sure to never accidentally hide a core plugin
+            return false;
+        }
+
+        $expiredPlugins = StaticContainer::get('Piwik\Plugins\Marketplace\Plugins\Expired');
+        $expiredPlugins = $expiredPlugins->getNamesOfExpiredPaidPlugins();
+
+        if (!empty($module) && in_array($module, $expiredPlugins, $strict = true)) {
+            if ($pluginManager->isValidPluginName($module) && Request::isApiRequest($_GET)) {
+                throw new Exception(Piwik::translate('Marketplace_PluginExpiredApiError', $module));
+            }
+
+            $module = 'Marketplace';
+            $action = 'expiredLicense';
+        }
     }
 
     public function getStylesheetFiles(&$stylesheets)
@@ -45,8 +73,12 @@ class Marketplace extends \Piwik\Plugin
 
     public static function isMarketplaceEnabled()
     {
-        return (bool) Config::getInstance()->Marketplace['enabled'] &&
-               Plugin\Manager::getInstance()->isPluginActivated('Marketplace');
+        return self::getPluginManager()->isPluginActivated('Marketplace');
+    }
+
+    private static function getPluginManager()
+    {
+        return Plugin\Manager::getInstance();
     }
 
 }