+++ CHANGELOG.md
@@ -5,7 +5,7 @@ This is a changelog for Piwik platform developers. All changes for our HTTP API'
 ## Piwik 2.16.1
 
 ### Internal change
- * The setting `[General]enable_marketplace=0/1` was moved to `[Marketplace]enabled=0/1`. The updater should automatically migrate an existing setting.
+ * The setting `[General]enable_marketplace=0/1` was removed, instead the new plugin Marketplace can be disabled/enabled. The updater should automatically migrate an existing setting.
 
 ## Piwik 2.16.0